% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer2vcfR.R
\name{assignGTfrombiWSNRAF}
\alias{assignGTfrombiWSNRAF}
\title{Assign the GT calls from the NONREFERENT WSAF}
\usage{
assignGTfrombiWSNRAF(wsnraf, cutoff = 0.1)
}
\value{
A GT matrix
not exported
}
\description{
Assign the GT calls from the NONREFERENT WSAF
}
