% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_MIPanalyzer_loaded}
\alias{check_MIPanalyzer_loaded}
\title{Check that MIPanalyzer package has loaded successfully}
\usage{
check_MIPanalyzer_loaded()
}
\description{
Simple function to check that MIPanalyzer package has loaded 
  successfully.
}
