% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_inbreeding_coefficient.R
\name{cluster_inbreeding_coef}
\alias{cluster_inbreeding_coef}
\title{Identify Inbreeding Coefficients of Locations in Continuous Space}
\usage{
cluster_inbreeding_coef(
  clst_gendist_geodist,
  start_params = c(),
  m_lowerbound = 0,
  m_upperbound = 1,
  learningrate = 0.05,
  steps = 100,
  report_progress = TRUE
)
}
\arguments{
\item{clst_gendist_geodist}{dataframe; The genetic-geographic ata}

\item{start_params}{named numeric vector; vector of start parameters.}

\item{m_lowerbound}{numeric; a lower bound for the m parameter}

\item{m_upperbound}{numeric; an upper bound for the m parameter}

\item{learningrate}{numeric; alpha parameter for how much each "step" is weighted in the gradient descent}

\item{steps}{numeric; the number of "steps" as we move down the gradient}

\item{report_progress}{boolean; whether or not a progress bar should be shown as you iterate through steps}
}
\description{
The purpose of this statistic is to identify an inbreeding coefficient, or degree of 
             relatedness, for a given location in space. We assume that locations in spaces can be 
             represented as "clusters," such that multiple individuals live in the same cluster 
             (i.e. samples are sourced from the same location). The expected pairwise relationship
             between two individuals, or samples, is dependent on the each sample's cluster's inbreeding
             coefficient and the geographic distance between the clusters.
}
\details{
The gen.geo.dist dataframe must be named with the following columns: 
         "smpl1"; "smpl2"; "locat1"; "locat2"; "gendist"; "geodist"; which corresponds to: 
         Sample 1 Name; Sample 2 Name; Sample 1 Location; Sample 2 Location; 
         Pairwise Genetic Distance; Pairwise Geographpic Distance. Note, the order of the 
         columns do not matter but the names of the columns must match.

The start_params vector names must match the cluster names (i.e. clusters must be 
         have a name that we can match on for the starting relatedness paramerts). In addition,
         you must provide a start parameter for "m".
}
