% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{explore_filter_coverage_loci}
\alias{explore_filter_coverage_loci}
\title{Explore locus coverage prior to filtering}
\usage{
explore_filter_coverage_loci(x, min_coverage = 5, max_missing = 50,
  breaks = 100)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{min_coverage}{any coverage value \code{< min_coverage} is replaced by \code{NA}.}

\item{max_missing}{any locus with \code{> max_missing} values equal \code{NA} are dropped.}

\item{breaks}{number of breaks spanning the range \code{[0,100]}.}
}
\description{
See \code{filter_coverage_loci()} function first. Explore values
  of \code{min_coverage} and \code{max_missing} without applying any
  filtering.
}
