% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter alleles based on raw counts}
\usage{
filter_counts(x, count_min = 2,
  description = "filter individual allele counts")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_multiallelic}.}

\item{count_min}{alleles with fewer than this many counts are dropped.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Drop any allele for which the number of read counts is below a
  given threshold. Coverage is adjusted to account for dropped reads.
}
