% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_coverage_samples}
\alias{filter_coverage_samples}
\title{Filter samples based on coverage}
\usage{
filter_coverage_samples(x, min_coverage = 5, max_missing = 50,
  description = "filter samples based on coverage")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{min_coverage}{any coverage value \code{< min_coverage} is replaced by \code{NA}.}

\item{max_missing}{any sample with \code{> max_missing} values equal \code{NA} are dropped.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Set a coverage threshold: any coverage value below this
  threshold is replaced by \code{NA}. Then set a maximum percent missing loci
  per sample: any sample with greater than this percentage missing loci is
  dropped. Note that threshold values can be explored without applying any
  filtering using the \code{explore_filter_coverage_samples()} function.
}
