% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_loci}
\alias{filter_loci}
\title{Filter out some loci}
\usage{
filter_loci(x, locus_filter, description = "")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic} or
\code{mipanalyzer_multiallelic}.}

\item{locus_filter}{boolean vector specifying whether to keep (\code{TRUE})
or drop (\code{FALSE}) each locus.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Filter out some loci.
}
