% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_loci_invariant}
\alias{filter_loci_invariant}
\title{Filter loci to drop invariant sites}
\usage{
filter_loci_invariant(x,
  description = "filter loci to drop invariant sites")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_multiallelic}.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Filter loci to drop invariant sites.
}
