% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_overcounts}
\alias{filter_overcounts}
\title{Filter out over-counts}
\usage{
filter_overcounts(x, description = "replace overcounts with NA")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic} or
\code{mipanalyzer_multiallelic}.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Filter out over-counts, defined as count > coverage.
  Replace any such element with NA.
}
