% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter out some samples}
\usage{
filter_samples(x, sample_filter, description = "")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{sample_filter}{boolean vector specifying whether to keep (\code{TRUE})
or drop (\code{FALSE}) each sample.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Filter out some samples.
}
