% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_wsaf}
\alias{filter_wsaf}
\title{Filter alleles based on within-sample allele frequencies}
\usage{
filter_wsaf(x, wsaf_min = 0.01, description = "filter individual allele WSAF")
}
\arguments{
\item{x}{object of class \code{mipanalyzer_multiallelic}.}

\item{wsaf_min}{alleles with counts that make a WSAF less than this threshold
are dropped.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Drop any allele for which the within-sample allele frequency
  (WSAF) is below a givin threshold. Thresholds apply in both directions, for
  example if \code{wsaf_min = 0.01} then alleles with a WSAF less than 0.01
  *or* greater than 0.99 will be rounded to 0 or 1, respectively. Coverage is
  adjusted to account for dropped reads.
}
