% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_IBS_distance}
\alias{get_IBS_distance}
\title{Get identity by state (IBS) distance}
\usage{
get_IBS_distance(x, ignore_het = TRUE, diagonal = NULL, report_progress = TRUE)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{ignore_het}{whether to ignore heterzygous comparisons, or alternatively
call the major allele at every locus (see details).}

\item{diagonal}{Should the diagonal of the distance matrix be changed to a
given value. Deafult = NULL, which cause no changes.}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console.}
}
\description{
Get identity by state (IBS) distance, computed as the proportion
  of sites that are identical between samples. If \code{ignore_het = TRUE}
  then heterozygous sites are ignored, otherwise the major strain is called
  at every locus.
}
