% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_wsaf}
\alias{get_wsaf}
\title{Get within-sample allele frequencies}
\usage{
get_wsaf(x, impute = TRUE, FUN = mean, ...)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{impute}{whether to impute missing values.}

\item{FUN}{function used to impute missing values. Default = `mean`}

\item{...}{other arguments to pass to \code{FUN}.}
}
\description{
Get within-sample allele frequencies from coverage and count
  data. Missing values can optionally be imputed by applying a summary
  function to the non NA values at each locus. The default summary function
  takes the mean of the non NA values.
}
