% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_biallelic.R
\name{is.mipanalyzer_biallelic}
\alias{is.mipanalyzer_biallelic}
\title{Determine if object is of class mipanalyzer_biallelic}
\usage{
is.mipanalyzer_biallelic(x)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}}
}
\description{
Determine if object is of class \code{mipanalyzer_biallelic}.
}
