% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_data.R
\name{is.mipanalyzer_data}
\alias{is.mipanalyzer_data}
\title{Determine if object is of class mipanalyzer_data}
\usage{
is.mipanalyzer_data(x)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_data}}
}
\description{
Determine if object is of class \code{mipanalyzer_data}.
}
