% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_multiallelic.R
\name{is.mipanalyzer_multiallelic}
\alias{is.mipanalyzer_multiallelic}
\title{Determine if object is of class mipanalyzer_multiallelic}
\usage{
is.mipanalyzer_multiallelic(x)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_multiallelic}}
}
\description{
Determine if object is of class \code{mipanalyzer_multiallelic}.
}
