% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIPanalyzer.R, R/package.R
\docType{package}
\name{MIPanalyzer}
\alias{MIPanalyzer}
\alias{MIPanalyzer-package}
\title{MIPanalyzer}
\description{
This package can be used to read in raw molecular inversion
  probe (MIP) data from vcf into a format that is convenient to work with.
  Data can be filtered based on counts, frequencies, missingness or other
  criteria. Filtered data can be analysed by common methods including PCA and
  various pairwise genetic metrics, and can be visualised in multiple ways.
  This package is intended to evolve as new MIP analyses are needed, thereby
  making it easy to repeat common analyses as new data becomes available.

Filtering and analysis of MIP data.
}
\author{
\strong{Maintainer}: Bob Verity \email{r.verity@imperial.ac.uk}

}
