% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{mipanalyzer_biallelic_to_vcfR}
\alias{mipanalyzer_biallelic_to_vcfR}
\title{Converts a MIPanalyzer biallelic object to vcfR}
\usage{
mipanalyzer_biallelic_to_vcfR(input = NULL, cutoff = 0.1)
}
\arguments{
\item{input}{an object of class \code{mipanalyzer_biallelic}.}

\item{cutoff}{the within-sample non-referent allele frequency cutoff to
transform your biallelic site to a genotype matrix.}
}
\description{
Converts an object of class \code{mipanalyzer_biallelic} to
  \code{vcfR} format.
}
