% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{mipanalyzer_file}
\alias{mipanalyzer_file}
\title{Load system file}
\usage{
mipanalyzer_file(name)
}
\arguments{
\item{name}{the name of a file within the inst/extdata folder.}
}
\description{
Load a file from within the inst/extdata folder of the
  MIPanalyzer package. File extension must be one of .csv, .txt, or .rds.
}
