% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer2vcfR.R
\name{MIPanalyzerbi2vcfR}
\alias{MIPanalyzerbi2vcfR}
\title{Convert a MIPanalyzer Biallelic Object to a vcfR object}
\usage{
MIPanalyzerbi2vcfR(input = NULL, cutoff = 0.1)
}
\arguments{
\item{input}{is of class \code{mipanalyzer_biallelic}}

\item{cutoff}{is the within-sample non-referent allele frequency cutoff to transform your biallelic site to a genotype matrix.}
}
\description{
Convert a MIPanalyzer Biallelic Object to a vcfR object
}
