% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pcoa_genomic_distance}
\alias{pcoa_genomic_distance}
\title{PCoA of genomic distances between samples}
\usage{
pcoa_genomic_distance(x)
}
\arguments{
\item{x}{matrix of genomic distances, as produced by the function
\code{get_genomic_distance()}.}
}
\description{
Conduct principal coordinate analysis (PCoA) on a matrix of
  genomic distances.
}
