% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage matrix}
\usage{
plot_coverage(x)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic} or
\code{mipanalyzer_multiallelic}.}
}
\description{
Plot matrix of coverage over all samples and loci.
}
