% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_distance.R
\name{plot_distance}
\alias{plot_distance}
\title{Plot a distance matrix}
\usage{
plot_distance(m, col_pal = "plasma")
}
\arguments{
\item{m}{square matrix of pairwise distances.}

\item{col_pal}{which viridis colour pallet to use. Options are "viridis",
"plasma", "magma" or "inferno".}
}
\description{
Simple image plot of a matrix of pairwise distances.
}
