% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_map}
\alias{plot_map}
\title{Produce ggplot map}
\usage{
plot_map(
  x_limits = c(12, 35),
  y_limits = c(-13, 5),
  col_country = grey(0.3),
  col_country_border = grey(0.5),
  size_country_border = 0.5,
  col_sea = grey(0.1),
  resolution = "coarse"
)
}
\arguments{
\item{x_limits}{longitude limits of map.}

\item{y_limits}{latitude limits of map.}

\item{col_country}{fill colour of countries.}

\item{col_country_border}{colour of country borders.}

\item{size_country_border}{size of country borders.}

\item{col_sea}{fill colour of sea.}

\item{resolution}{the resolution of the underlying map. Must be one of
"coarse", "low", "less", "islands", "li", "high".}
}
\description{
Produce ggplot map.
}
