% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_clustering.R
\name{plot_pca_contribution}
\alias{plot_pca_contribution}
\title{Plot PCA contribution of each variable}
\usage{
plot_pca_contribution(
  pca,
  component = 1,
  chrom,
  pos,
  locus_type = NULL,
  y_buffer = 0
)
}
\arguments{
\item{pca}{output of \code{pca_wsaf()} function.}

\item{component}{which component to plot.}

\item{chrom}{the chromosome corresponding to each contribution value.}

\item{pos}{the genomic position corresponding each contribution value.}

\item{locus_type}{defines the colour of each bar.}

\item{y_buffer}{(percent). A buffer added to the bottom of each y-axis,
making room for other annotations to be added.}
}
\description{
Plot PCA contribution of each variable.
}
