% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_pca_loadings}
\alias{plot_pca_loadings}
\title{Plot PCA loadings}
\usage{
plot_pca_loadings(pca, component = 1, chrom, pos, locus_type = NULL,
  y_buffer = 0)
}
\arguments{
\item{pca}{output of \code{pca_wsaf()} function.}

\item{component}{which component to plot.}

\item{chrom}{the chromosome for each loading value.}

\item{pos}{the genomic position for each loading value.}

\item{locus_type}{defines the colour of each bar.}

\item{y_buffer}{(percent). A buffer added to the bottom of each y-axis,
making room for other annotations to be added.}
}
\description{
Plot loading values of PCA against genomic position.
}
