% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_pca_variance}
\alias{plot_pca_variance}
\title{Plot variance explained by PCA components}
\usage{
plot_pca_variance(pca, num_components = 10)
}
\arguments{
\item{pca}{output of \code{pca_wsaf()} function.}

\item{num_components}{maximum components to be shown.}
}
\description{
Plot the variance explained by each PCA component. The number of
  components shown is controlled by \code{num_components}, with up to the
  first 10 componenets shown by default. If less than the requested number of
  components exist, then all the components will be shown.
}
