% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_clustering.R
\name{plot_pcoa}
\alias{plot_pcoa}
\title{Plot PCoA}
\usage{
plot_pcoa(pcoa, num_components = 2, col = NULL, col_palette = NULL)
}
\arguments{
\item{pcoa}{object of class "pcoa", as produced by \code{pcoa_wsaf()} function.}

\item{num_components}{numeric for number of components used. Default = 2.}

\item{col}{vector by which samples are coloured.}

\item{col_palette}{vector of colours for each group.}
}
\description{
Plots either the first 2 or 3 vectors of PCoA.
}
