% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_wsaf}
\alias{plot_wsaf}
\title{Plot within-sample allele frequencies}
\usage{
plot_wsaf(x, col_pal = "plasma")
}
\arguments{
\item{x}{matrix of within-sample allele frequencies, with samples in rows and
loci in columns.}

\item{col_pal}{which viridis colour pallet to use. Options are "viridis",
"plasma", "magma" or "inferno".}
}
\description{
Simple image plot of within-sample allele frequencies. The top
  row of the plot corresponds to the first row of the input matrix.
}
