% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_biallelic.R
\name{print.mipanalyzer_biallelic}
\alias{print.mipanalyzer_biallelic}
\title{Custom print function for class mipanalyzer_biallelic}
\usage{
\method{print}{mipanalyzer_biallelic}(x, ...)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class \code{mipanalyzer_biallelic},
  printing a summary of the key elements (also equivalent to
  \code{summary(x)}). To do an ordinary \code{print()}, use the
  \code{print_full()} function.
}
