% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_data.R
\name{print.mipanalyzer_data}
\alias{print.mipanalyzer_data}
\title{Custom print function for class mipanalyzer_data}
\usage{
\method{print}{mipanalyzer_data}(x, ...)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_data}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class \code{mipanalyzer_data},
  printing a summary of the key elements (also equivalent to
  \code{summary(x)}). To do an ordinary \code{print()}, use the
  \code{print_full()} function.
}
