% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_multiallelic.R
\name{print.mipanalyzer_multiallelic}
\alias{print.mipanalyzer_multiallelic}
\title{Custom print function for class mipanalyzer_multiallelic}
\usage{
\method{print}{mipanalyzer_multiallelic}(x, ...)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_multiallelic}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class \code{mipanalyzer_multiallelic},
  printing a summary of the key elements (also equivalent to
  \code{summary(x)}). To do an ordinary \code{print()}, use the
  \code{print_full()} function.
}
