% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_biallelic}
\alias{sim_biallelic}
\title{Simulate biallelic data}
\usage{
sim_biallelic(COI = 3, PLAF = seq(0, 0.5, 0.01), coverage = 100,
  alpha = 1, overdispersion = 0, epsilon = 0)
}
\arguments{
\item{COI}{complexity of infection.}

\item{PLAF}{vector of population-level allele frequencies.}

\item{coverage}{coverage at each locus. If a single value then the same
coverage is applied over all loci.}

\item{alpha}{shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{the extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-binomially distributed, with
the beta distribution having shape parameters \code{p/overdispersion} and
\code{(1-p)/overdispersion}.}

\item{epsilon}{the probability of a single read being mis-called as the other
allele. Applies in both directions.}
}
\description{
Simulate biallelic data.
}
\details{
TODO
}
