% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_biallelic.R
\name{summary.mipanalyzer_biallelic}
\alias{summary.mipanalyzer_biallelic}
\title{Print summary for class mipanalyzer_biallelic}
\usage{
\method{summary}{mipanalyzer_biallelic}(object, ...)
}
\arguments{
\item{object}{object of class \code{mipanalyzer_biallelic}}

\item{...}{other arguments (ignored)}
}
\description{
Custom summary function for class \code{mipanalyzer_biallelic}.
}
