% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_data.R
\name{summary.mipanalyzer_data}
\alias{summary.mipanalyzer_data}
\title{Print summary for class mipanalyzer_data}
\usage{
\method{summary}{mipanalyzer_data}(object, ...)
}
\arguments{
\item{object}{object of class \code{mipanalyzer_data}}

\item{...}{other arguments (ignored)}
}
\description{
Custom summary function for class \code{mipanalyzer_data}.
}
