% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer_multiallelic.R
\name{summary.mipanalyzer_multiallelic}
\alias{summary.mipanalyzer_multiallelic}
\title{Print summary for class mipanalyzer_multiallelic}
\usage{
\method{summary}{mipanalyzer_multiallelic}(object, ...)
}
\arguments{
\item{object}{object of class \code{mipanalyzer_multiallelic}}

\item{...}{other arguments (ignored)}
}
\description{
Custom summary function for class \code{mipanalyzer_multiallelic}.
}
