% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vcf2mipanalyzer_multiallelic}
\alias{vcf2mipanalyzer_multiallelic}
\title{Convert vcf to multiallelic mipanalyzer data class}
\usage{
vcf2mipanalyzer_multiallelic(file = NULL, vcfR = NULL, verbose = TRUE)
}
\arguments{
\item{file}{path to vcf file.}

\item{vcfR}{object of class \code{vcfR}.}

\item{verbose}{if reading from file, whether to read in verbose manner.}
}
\description{
Convert vcf to multiallelic mipanalyzer data class.
}
