% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipanalyzer2vcfR.R
\name{assignGTfrombiWSNRAF}
\alias{assignGTfrombiWSNRAF}
\title{Assign the GT calls from the NONREFERENT WSAF}
\usage{
assignGTfrombiWSNRAF(wsnraf, cutoff = 0.1)
}
\arguments{
\item{wsnraf}{A within-sample non-referent allele frequncy from biallelic SNPs as a \code{matrix}.}

\item{cutoff}{The allele-frequency cutoff to determine genotype calls}
}
\value{
A GT matrix as a \code{matrix}.
not exported
}
\description{
This function takes in the within-sample non-referent allele frequency matrix and converts it to a (diploid) genotype matrix. The genotype matrix is 
character matrix with "0/0", "0/1", "1/1" representing the homozygote referent, heterozygote, and homozygote alternative calls. The genotype call is 
determined by the allele frequency and the \code{cutoff} set by the user. Specifically, an allele frequency less than the cutoff or greater than \code{1-cutoff} correspond to a
homozygote referent and homozygote alternative call, respectively. All other allele-frequencies will be converted to the heterozygote call \code{"0/1"}.
}
