% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{explore_filter_coverage_loci}
\alias{explore_filter_coverage_loci}
\title{Explore locus coverage prior to filtering}
\usage{
explore_filter_coverage_loci(x, min_coverage = 5,
  max_low_coverage = 50, breaks = 100)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{min_coverage}{the coverage threshold below which data is deemed to be
low-coverage.}

\item{max_low_coverage}{(percentage). Loci are not allowed to contain more
than this many low-coverage samples. In the \code{filter_coverage_loci()}
function, any locus with more than \code{max_low_coverage} low-coverage
samples will be dropped.}

\item{breaks}{number of breaks spanning the range \code{[0,100]}.}
}
\description{
Explore what effect the \code{filter_coverage_loci()}
  function will have on the data without actually applying any filters. Can
  be used to set coverage thresholds.
}
