% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_coverage_loci}
\alias{filter_coverage_loci}
\title{Filter loci based on coverage}
\usage{
filter_coverage_loci(
  x,
  min_coverage = 5,
  max_low_coverage = 50,
  replace_low_coverage = FALSE,
  description = "filter loci based on coverage"
)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic} or
\code{mipanalyzer_multiallelic}.}

\item{min_coverage}{the coverage threshold below which data is deemed to be
low-coverage.}

\item{max_low_coverage}{any locus with more than \code{max_low_coverage}
percent of low-coverage samples will be dropped.}

\item{replace_low_coverage}{(Boolean). If \code{TRUE} then any remaining
low-coverage loci will be replaced with \code{NA}.}

\item{description}{brief description of the filter, to be saved in the filter
history.}
}
\description{
Set a coverage threshold: any coverage value below this
  threshold is deemed to be low-coverage. Then set a maximum percent
  low-coverage samples per locus: any locus with greater than this percentage
  low-coverage samples is dropped. Note that threshold values can be explored
  without applying any filtering using the
  \code{explore_filter_coverage_loci()} function.
}
