% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_IB_mixture}
\alias{get_IB_mixture}
\title{Get identity by mixture}
\usage{
get_IB_mixture(x, tol = 0, diagonal = NULL, report_progress = TRUE)
}
\arguments{
\item{x}{object of class \code{mipanalyzer_biallelic}.}

\item{tol}{tolerance on mixture comparisons. Defeault = 0}

\item{diagonal}{Should the diagonal of the distance matrix be changed to a
given value. Deafult = NULL, which cause no changes.}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console.}
}
\description{
Get identity by mixture distance.
}
