% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pca_wsaf}
\alias{pca_wsaf}
\title{PCA of within-sample allele frequencies}
\usage{
pca_wsaf(x)
}
\arguments{
\item{x}{a matrix of within-sample allele frequencies, as produced by the
function \code{get_wsaf()}.}
}
\value{
Invisibly returns a list of class `prcomp` with the following 
  components
  \itemize{
      \item{"sdev"}{ the standard deviations of the principal components
      (i.e., the square roots of the eigenvalues of the
      covariance/correlation matrix, though the calculation is actually done
      with the singular values of the data matrix).} \item{"rotation"}{ the
      matrix of variable loadings (i.e., a matrix whose columns contain the
      eigenvectors). The function \code{princomp} returns this in the element
      \code{loadings}.} \item{"center, scale"}{ the centering and scaling
      used.} \item{"x"}{ the value of the rotated data (the centred data
      multiplied by the rotation matrix). Hence, \code{cov(x)} is the
      diagonal matrix \code{diag(sdev^2)}.}
      \item{"var"}{ the variance in the data explained by each component.}
      \item{"contribution"}{ the percent contribution of a variable (i.e. a
      locus) to the overall variation.}
      }
}
\description{
Conduct principal components analysis (PCA) on a matrix of
  within-sample allele frequencies (WSAF). Missing values must have been
  already imputed. Output includes the raw components, the variance in the
  data explained by each component, and the loadings of each component also
  returned.
}
\details{
Contributions of each variable are computed from the loading values
  (stored as "rotation" within the \code{prcomp} object). The percent
  contribution of a variable is defined as the absolute loading value for
  this variable, divided by the sum of loadings over all variables and
  multiplied by 100.
}
